0.11 fix php tests (#522)
* Restructuring and adding bib tests

* fix user api tests

* add code coverage info to readme

* Fixed Api Editor Controller partially

* fix basic api tests

* Finished ApiEditorTests

* started fixing entity tests

* fix search api tests

* add thesaurus permissions to seeder

* add testdox

* fix preference api tests

* fix notification and comment tests

* fix order; indent testdox http verbs

* fix unit tests

* Fixed ApiEntityTests, added ApiRefTest and Fixed some Entitiy issues

* update phpunit config

* add vite to phpunit ci

* fix spaces and comments; fix wrong usage of `

* Added comment and reference tests and dome fixes.

* Removed File and Map Tests

* Adjustes user tests for new user seeder

* Fixed activity test.

* fixed commentcount

* removed map references

* Fixed tests

* Fixed all tests

* Renamed PermissionTester to ResponseTester

* restructure seeders

* fix missing token refresh in tests

* disable map related code

* simplify save on user patch

---------

+++ app/Http/Controllers/UserController.php
@@ -45,7 +45,7 @@ class UserController extends Controller
                     case 'App\Entity':
                         try {
                             $name = Entity::findOrFail($n->data['resource']['id'])->name;
-                        } catch (ModelNotFoundException $e) {
+                        } catch(ModelNotFoundException $e) {
                             $skip = true;
                         }
                         break;
@@ -54,7 +54,7 @@ class UserController extends Controller
                         try {
                             $name = Entity::findOrFail($n->data['resource']['meta']['entity_id'])->name;
                             $attrUrl = Attribute::findOrFail($n->data['resource']['meta']['attribute_id'])->thesaurus_url;
-                        } catch (ModelNotFoundException $e) {
+                        } catch(ModelNotFoundException $e) {
                             $skip = true;
                         }
                         break;
@@ -69,7 +69,7 @@ class UserController extends Controller
                     if(isset($attrUrl)) {
                         $data['attribute_url'] = $attrUrl;
                     }
-                    
+
                     $n->info = $data;
                 }
             } else if($n->type == 'App\Notifications\EntityUpdated') {
@@ -77,7 +77,7 @@ class UserController extends Controller
                     $n->info = [
                         'name' => Entity::findOrFail($n->data['resource']['id'])->name,
                     ];
-                } catch (ModelNotFoundException $e) {
+                } catch(ModelNotFoundException $e) {
                 }
             }
             return $n;
@@ -335,24 +335,31 @@ class UserController extends Controller
             // Update updated_at column
             $user->touch();
         }
+        $saveRequired = false;
         if($request->has('email')) {
             $user->email = Str::lower($request->get('email'));
-            $user->save();
+            $saveRequired = true;
         }
         if($request->has('name')) {
             $user->name = $request->get('name');
-            $user->save();
+            $saveRequired = true;
         }
         if($request->has('nickname')) {
             $user->nickname = Str::lower($request->get('nickname'));
-            $user->save();
+            $saveRequired = true;
         }
-        $user->setMetadata(
-            $request->only('phonenumber', 'orcid', 'role', 'field', 'institution', 'department')
-        );
+        $metadataFields = $request->only('phonenumber', 'orcid', 'role', 'field', 'institution', 'department');
+        if(count($metadataFields) > 0) {
+            $user->setMetadata($metadataFields);
+            $saveRequired = true;
+        }
 
         // return user without roles relation
         $user->unsetRelation('roles');
+
+        if($saveRequired) {
+            $user->save();
+        }
 
         return response()->json($user);
     }
@@ -360,7 +367,7 @@ class UserController extends Controller
     public function restoreUser($id)
     {
         $user = auth()->user();
-        if (!$user->can('users_roles_delete')) {
+        if(!$user->can('users_roles_delete')) {
             return response()->json([
                 'error' => __('You do not have the permission to restore users')
             ], 403);
@@ -368,7 +375,7 @@ class UserController extends Controller
 
         try {
             $delUser = User::onlyTrashed()->findOrFail($id);
-        } catch (ModelNotFoundException $e) {
+        } catch(ModelNotFoundException $e) {
             return response()->json([
                 'error' => __('This user does not exist')
             ], 400);
@@ -422,6 +429,7 @@ class UserController extends Controller
             $role->description = $request->get('description');
         }
         $role->save();
+        $role->permissions;
 
         return response()->json($role);
     }
@@ -483,7 +491,7 @@ class UserController extends Controller
             $password = Hash::make($request->get('password'));
             $pUser->password = $password;
         }
-        
+
         $pUser->login_attempts = null;
         $pUser->save();
 
@@ -545,7 +553,9 @@ class UserController extends Controller
             ], 400);
         }
 
-        Storage::delete($user->avatar);
+        if(isset($user->avatar)) {
+            Storage::delete($user->avatar);
+        }
         $user->avatar = null;
         $user->save();
         return response()->json(null, 204);
